fix ActorControlSelfDelegate, updated NuGet package dependencies

+++ NecroLens/Service/DeepDungeonService.cs
@@ -33,7 +33,7 @@ public class DeepDungeonService : IDisposable
     public readonly Dictionary<Pomander, string> PomanderNames;
     
     private const string ActorControlSig = "E8 ?? ?? ?? ?? 0F B7 0B 83 E9 64";
-    private delegate void ActorControlSelfDelegate(uint category, uint eventId, uint param1, uint param2, uint param3, ulong param4, uint param5, uint param6, uint targetId, uint param7, uint param8, uint param9);
+    private delegate void ActorControlSelfDelegate(uint category, uint eventId, uint param1, uint param2, uint param3, uint param4, uint param5, uint param6, uint targetId, uint param7, ulong param8, byte param9);
     private Hook<ActorControlSelfDelegate>? actorControlSelfHook;
     
     
@@ -138,7 +138,7 @@ public class DeepDungeonService : IDisposable
         FloorTimes[FloorDetails.CurrentFloor] = time;
     }
 
-    private void ActorControlSelf(uint category, uint eventId, uint param1, uint param2, uint param3, ulong param4, uint param5, uint param6, uint targetId, uint param7, uint param8, uint param9)
+    private void ActorControlSelf(uint category, uint eventId, uint param1, uint param2, uint param3, uint param4, uint param5, uint param6, uint targetId, uint param7, ulong param8, byte param9)
     {
         
         actorControlSelfHook!.Original(category, eventId, param1, param2, param3, param4, param5, param6, targetId, param7, param8, param9);
